Fix incorrect CHS values displayed in the new hard disk dialog dropdown

+++ src/qt/qt_harddiskdialog.cpp
@@ -73,7 +73,7 @@ HarddiskDialog::HarddiskDialog(bool existing, QWidget *parent) :
     model = ui->comboBoxType->model();
     for (int i = 0; i < 127; i++) {
         uint64_t size = ((uint64_t) hdd_table[i][0]) * hdd_table[i][1] * hdd_table[i][2];
-        uint64_t size_mb = size >> 11LL;
+        uint32_t size_mb = size >> 11LL;
         //QString text = QString("%1 MiB (CHS: %2, %3, %4)").arg(size_mb).arg(hdd_table[i][0]).arg(hdd_table[i][1]).arg(hdd_table[i][2]);
         QString text = QString::asprintf(tr("%u MB (CHS: %i, %i, %i)").toUtf8().constData(), (size_mb), (hdd_table[i][0]), (hdd_table[i][1]), (hdd_table[i][2]));
         Models::AddEntry(model, text, i);
